moved provider to correct module
